simplify things and get rid of Ensemble and Conformation classes. Now, Specimen optionally has a conformation to simulate an image from a stacked ElectronDensity
