+++ src/cryojax/simulator/density/_atom_density.py
@@ -63,6 +63,6 @@ class AtomCloud(ElectronDensity):
         """
         raise NotImplementedError
 
-    def __getitem__(self, key: int):
+    def __getitem__(self, key: int) -> "AtomCloud":
         """Get a particular electron density in the electron density stack."""
         raise NotImplementedError
