--- src/cryojax/simulator/specimen/__init__.py
@@ -1,6 +1,0 @@
-from . import _specimen, _ensemble
-
-from ._specimen import *
-from ._ensemble import *
-
-__all__ = _specimen.__all__ + _ensemble.__all__
